Convert TreeTables functions to components
