+++ assets/src/scripts/components/Pipes.jsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+function Pipes({ pipes }) {
+  return pipes.map((pipe, ix) => (
+    <span
+      key={ix}
+      style={{
+        display: "inline-block",
+        textAlign: "center",
+        paddingLeft: "3px",
+        paddingRight: "6px",
+        width: "20px",
+      }}
+    >
+      {pipe}
+    </span>
+  ));
+}
+
+export default Pipes;
